requirements.txt
numpy
streamlit
seaborn
matplotli
plotly
panda
joblib
scikit-learn